pyproject.toml
[project]
name = "cpr-dataanalysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "marimo>=0.13.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "plotly>=5.0",
    "pyworms>=0.1",
    "scipy>=1.11",
    "matplotlib>=3.10.8",
    "lda4microbiome>=0.1.21",
    "kaleido>=1.2.0",
    "umap-learn>=0.5.11",
    "multi-mst>=0.2.0",
    "fast-hdbscan>=0.2.0,<0.3.0",
    "bokeh>=3.9.0",
    "colorcet>=3.1.0",
    "holoviews>=1.22.1",
    "scikit-image>=0.26.0",
]
